# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  response.Response:
    properties:
      data: {}
      error: {}
      message:
        type: string
      success:
        type: boolean
    type: object
...
  serializers.MerchantListResponse:
    properties:
      merchants:
        items:
          $ref: '#/definitions/serializers.MerchantResponse'
        type: array
      page:
        type: integer
      page_size:
        type: integer
      total:
        type: integer
    type: object
  serializers.MerchantResponse:
    properties:
      created_at:
        type: string
      id:
        type: integer
      merchant_name:
        type: string
      payment_merchant_id:
        type: string
      payment_provider_id:
        type: integer
      review_status:
        type: integer
      shop_id:
        type: integer
      shop_url:
        type: string
      updated_at:
        type: string
    type: object
...
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: Payment API for Makeshop
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Makeshop Payment API
  version: "1.0"
paths:
  /admin/audit-logs:
    get:
      consumes:
      - application/json
      description: Get a list of all audit logs in the system with optional filtering
      parameters:
      - description: 'Page number (default: 1)'
        in: query
        name: page
        type: integer
      - description: 'Page size (default: 10)'
        in: query
        name: page_size
        type: integer
      - description: 'Field to sort by (default: id)'
        in: query
        name: sort_field
        type: string
      - description: 'Sort order: ascend or descend (default: ascend)'
        in: query
        name: sort_order
        type: string
      - description: Filter by created date (RFC3339, RFC1123, or YYYY-MM-DD format)
        in: query
        name: created_at
        type: string
      - description: Filter by user ID
        in: query
        name: user_id
        type: integer
      - description: Filter by keyword in description
        in: query
        name: keyword
        type: string
      - description: Filter by audit log type
        in: query
        name: audit_log_type
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: List all audit logs
      tags:
      - Admin Audit Log Management
  /admin/audit-logs/users:
    get:
      consumes:
      - application/json
      description: Retrieve a list of distinct users who have audit log entries
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Get distinct users with audit logs
      tags:
      - Admin Audit Log Management
  /admin/merchants:
    get:
      consumes:
      - application/json
      description: Get a list of merchants with optional filtering and pagination
      parameters:
      - description: 'Page number (default: 1)'
        in: query
        name: page
        type: integer
      - description: 'Page size (default: 10)'
        in: query
        name: page_size
        type: integer
      - description: Search query for merchant name, ID, etc.
        in: query
        name: search
        type: string
      - description: 'Filter by review status (values 1-3, can use multiple times:
          review_status=1&review_status=2)'
        in: query
        name: review_status
        type: array
      - description: Filter by creation date start (RFC3339 format)
        in: query
        name: created_at_start
        type: string
      - description: Filter by creation date end (RFC3339 format)
        in: query
        name: created_at_end
        type: string
      - description: Field to sort by (e.g., created_at, merchant_name, etc.)
        in: query
        name: sort_field
        type: string
      - description: 'Sort order (asc or desc, default: desc)'
        in: query
        name: sort_order
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/serializers.MerchantListResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: List merchants
      tags:
      - Merchant Management
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"